Refactor: 상품 엔티티 인기 점수 업데이트 메서드 타입 변경
Former-commit-id: 4dfc497fcbc4636fe100da2747f83cb358c5aa41

+++ back/src/main/java/site/chachacha/fitme/domain/product/entity/Product.java
@@ -115,7 +115,7 @@ public class Product extends BaseEntity {
         this.likeCount = this.likeCount + 1;
     }
 
-    public void updateMonthlyPopularityScore(int monthlyPopularityScore) {
+    public void updateMonthlyPopularityScore(double monthlyPopularityScore) {
         this.monthlyPopularityScore = monthlyPopularityScore;
     }
